requirements/backend/jwt2fa/requirements.txt
Django>=4.0,<6.0
djangorestframework>=3.14
psycopg2-binary>=2.9,<3.0
python-dotenv>=0.21
requests>=2.25.1
gunicorn==20.1.0
django-cors-headers
sendgrid-django
sendgrid
pillow
dnspython
djangorestframework-simplejwt
